third commit - delete some non essential parts of architecture

+++ src/main/java/fantastic_stories_app/repository/AuthorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorRepository extends PagingAndSortingRepository<Author, Long> {
+public interface AuthorRepository extends PagingAndSortingRepository<Author, Integer> {
 }
