+++ src/main/java/fantastic_stories_app/repository/IssueRepository.java
@@ -3,5 +3,5 @@ package fantastic_stories_app.repository;
 import fantastic_stories_app.model.Issue;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface IssueRepository extends PagingAndSortingRepository<Issue, Long> {
+public interface IssueRepository extends PagingAndSortingRepository<Issue, Integer> {
 }
