+++ src/main/java/fantastic_stories_app/repository/ReviewRepository.java
@@ -3,5 +3,5 @@ package fantastic_stories_app.repository;
 import fantastic_stories_app.model.Review;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ReviewRepository extends PagingAndSortingRepository<Review, Long> {
+public interface ReviewRepository extends PagingAndSortingRepository<Review, Integer> {
 }
