+++ src/main/java/fantastic_stories_app/repository/StoryRepository.java
@@ -3,5 +3,5 @@ package fantastic_stories_app.repository;
 import fantastic_stories_app.model.Story;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface StoryRepository extends PagingAndSortingRepository<Story, Long> {
+public interface StoryRepository extends PagingAndSortingRepository<Story, Integer> {
 }
